Script updated to handle few exceptional cases.
Script was getting failed in a scenario when a user of the group who writes a post, later decides to leave the group or is removed from the group. This case was leading to null value being scraped for the user id of the person who wrote that post. This null value was causing the script to fail. This has been fixed now.